Keeping CDash happy
git-svn-id: https://svn.opendtect.org/od/trunk@31971 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/Basic/survinfo.cc
@@ -64,13 +64,13 @@ TrcKey InlCrlSystem::nearestTrace( const Coord& crd, float* dist ) const
 	if ( cs_.hrg.includes(tk.pos()) )
 	{
 	    const Coord projcoord( transform(tk.pos()) );
-	    *dist = projcoord.distTo( crd );
+	    *dist = (float)projcoord.distTo( crd );
 	}
 	else
 	{
 	    BinID nearbid( cs_.hrg.getNearest(tk.pos()) );
 	    const Coord nearcoord( transform(nearbid) );
-	    *dist = nearcoord.distTo( crd );
+	    *dist = (float)nearcoord.distTo( crd );
 	}
     }
     return tk;
